% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_isotype_usage.R
\name{barplot_isotype_usage}
\alias{barplot_isotype_usage}
\title{Barplot isotype usage}
\usage{
barplot_isotype_usage(
  list_repertoires,
  order_ids,
  chain = "H",
  sample_name,
  text_size = 5,
  width_height = c(x = 10, y = 10),
  figure_path = "",
  color_scale_category = c("#66c2a5", "#fc8d62", "#8da0cb", "#e78ac3", "#a6d854",
    "#ffd92f")
)
}
\arguments{
\item{list_repertoires}{Patient Metadata.}

\item{figure_path}{Path to which figures should be saved.}
}
\value{
TRUE (plots saved as pdfs into figure_path)
}
\description{
Barplot isotype usage
}
\examples{
plot_isotype_usage_bar(list_repertoires=example_data, figure_path="")
}
