% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_v_identity.R
\name{boxplot_v_identity}
\alias{boxplot_v_identity}
\title{Boxplot for comparison of V_identity across samples}
\usage{
boxplot_v_identity(
  list_repertoires,
  rep_category,
  chain = "H",
  color_scale = c("#d7191c", "#fdae61", "#2c7bb6", "#ffffbf", "#abd9e9"),
  plot_name = "",
  figure_path = ""
)
}
\arguments{
\item{list_repertoires}{List of repertoires}

\item{rep_category}{metadata.}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Boxplot for comparison of V_identity across samples
}
\examples{
boxplot_v_identity(list_repertoires=example_data,rep_category=sample_overview, path_figures="")
}
