% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_length_across_reps.R
\name{cdr_length_across_reps}
\alias{cdr_length_across_reps}
\title{Plot comparing CDR3 lengths across repertoies}
\usage{
cdr_length_across_reps(
  list_repertoires,
  rep_category,
  length_range = c(6:28),
  chain = "H",
  unique_seqs = TRUE,
  size_x_axis = 10,
  width_height = c(x = 15, y = 15),
  color_scale,
  sample_name = "",
  figure_path = ""
)
}
\arguments{
\item{list_repertoires}{List of repertoires}

\item{rep_category}{Metadata}

\item{length_range}{CDR lengths to be considered}

\item{chain}{Heavy (H) or Light (L)}

\item{unique_seqs}{If true sequences are uniqued before length dist is calculated}

\item{size_x_axis}{Size of text on x-axis}

\item{width_height}{Width and height of plot}

\item{color_scale}{Color scale to be applied}

\item{sample_name}{ID of analysis for plot naming}

\item{figure_path}{Path to which figure is saved}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Plot comparing CDR3 lengths across repertoies
}
\examples{
cdr_length_across_reps(list_repertoires, rep_category, length_range=c(6:28),chain="H",unique_seqs=TRUE, size_x_axis=10,width_height=c(x=15,y=15), color_scale, sample_name="",figure_path="")

}
