% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_expansion_rank.R
\name{clonal_expansion_rank}
\alias{clonal_expansion_rank}
\title{Faceted barplot showing isotype and frequency of top 10 expanded clones per sample}
\usage{
clonal_expansion_rank(
  dataset_summary,
  color_scale = c("#edf8fb", "#b3cde3", "#8c96c6", "#88419d", "black"),
  figure_path
)
}
\arguments{
\item{dataset_summary}{Data of sequneces and clone count/freq}

\item{color_scale}{Color scale for isotypes}

\item{path_figures}{Path to which figures should be saved.}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Faceted barplot showing isotype and frequency of top 10 expanded clones per sample
}
\examples{
clonal_expansion_rank(dataset_summary,,color_scale,path_figures="./figures")

}
