% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_heatmap_geom_tile.R
\name{enrichment_heatmap_geom_tile}
\alias{enrichment_heatmap_geom_tile}
\title{Heatmap of mAb enrichment using geom_tile}
\usage{
enrichment_heatmap_geom_tile(
  compare_rep_names = c("nCoV36_S1", "nCoV36_S2"),
  reference_rep_name = "nCoV36_Ab(Background)",
  ref_seqs_name = "ref_n36",
  names_cohort = "default",
  ordered_by = "S2",
  text_size = c(5, 5),
  figure_path
)
}
\arguments{
\item{compare_rep_names}{2d vector containing names of samples}

\item{reference_rep_name}{Name of dataset to compare against}

\item{ref_seqs_name}{Reference file (background detection)}

\item{names_cohort}{ID to add to filename}

\item{ordered_by}{Entry by which heatmap should be ordered}

\item{figure_path}{Path to which figures are saved}

\item{mode}{Way in which analysis is performed (pairwise or multi)}
}
\value{
TRUE
#' @examples
enrichment_heatmap_geom_tile(compare_rep_names=c("nCoV36_S1","nCoV36_S2"),reference_rep_name="nCoV36_Ab(Background)",ordered_by="S2",ref_seqs_name="ref_n36",names_cohort="default",mode="multi",figure_path)
}
\description{
Heatmap of mAb enrichment using geom_tile
}
