% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavy_light_comparisons.R
\name{heavy_light_comparisons}
\alias{heavy_light_comparisons}
\title{PLots comparing heavy and light chain usage}
\usage{
heavy_light_comparisons(
  paired_df = paired_df_example,
  subset = "all",
  figure_path
)
}
\arguments{
\item{paired_df}{Dataframe with paired seqdata}

\item{subset}{ID for naming}

\item{figure_path}{Path to which figures should be saved.}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
PLots comparing heavy and light chain usage
}
\examples{
heavy_light_comparisons(paired_df=paired_df_example,subset="all",figure_path)

}
