% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_figure.R
\name{network_figure}
\alias{network_figure}
\title{Network plot of binders and closest found neighbors}
\usage{
network_figure(binders_all, binder_matches, figure_path = "")
}
\arguments{
\item{binders_all}{Identified binders}

\item{binder_matches}{Closest neighbors to identified binders from repertoire data.}

\item{figure_path}{Path where figures are saved}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Network plot of binders and closest found neighbors
}
\examples{
network_figure(binders_all,binder_matches,figure_path="")

}
