% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_isotype_usage_bar.R
\name{plot_isotype_usage_bar}
\alias{plot_isotype_usage_bar}
\title{Barplot isotype usage}
\usage{
plot_isotype_usage_bar(
  list_repertoires,
  order_ids,
  chain = "H",
  sample_name,
  text_size = 5,
  width_height = c(x = 10, y = 10),
  path_figures = "",
  color_scale_category = c("#66c2a5", "#fc8d62", "#8da0cb", "#e78ac3", "#a6d854",
    "#ffd92f")
)
}
\arguments{
\item{list_repertoires}{Patient Metadata.}

\item{path_figures}{Path to which figures should be saved.}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Barplot isotype usage
}
\examples{
plot_isotype_usage_bar(list_repertoires=example_data, path_figures="")
}
