% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patients.R
\name{plot_patients}
\alias{plot_patients}
\title{Sideways barplot for patient timeline}
\usage{
plot_patients(patients = patients_metadata, figure_path = "Ehling_figs")
}
\arguments{
\item{patients}{Patient Metadata.}

\item{path_figures}{Path to which figures should be saved.}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Sideways barplot for patient timeline
}
\examples{
plot_patients(patients=patients_metadata,path_figures="./figures")
}
