% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_pool_analysis_132.R
\name{supp_pool_analysis_132}
\alias{supp_pool_analysis_132}
\title{Generates supplementary plots characterizing the 132 chosen sequences}
\usage{
supp_pool_analysis_132(chosen_132, figure_path)
}
\arguments{
\item{chosen_132}{Dataframe of 132 chosen sequences (96+36)}

\item{figure_path}{Path where figures are saved}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Generates supplementary plots characterizing the 132 chosen sequences
}
\examples{
supp_pool_analysis_132(chosen_132,binder_matches,figure_path="")

}
