% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vj_across_reps.R
\name{vj_across_reps}
\alias{vj_across_reps}
\title{Plot comparing V,J usage across repertoires}
\usage{
vj_across_reps(
  list_repertoires,
  rep_category,
  chain = "H",
  freq_threshold = 0.01,
  unique_rows = TRUE,
  size_x_axis = 5,
  size_y_axis = 5,
  figure_path,
  plot_name = "",
  color_scale
)
}
\arguments{
\item{list_repertoires}{List of repertoires}

\item{rep_category}{Metadata}

\item{chain}{Heavy (H) or Light (L)}

\item{freq_threshold}{threshold above which V,J genes are plotted}

\item{unique_rows}{If true all duplicated rows are excluded}

\item{size_x_axis}{Size of text on x-axis}

\item{size_y_axis}{Size of text on y-axis}

\item{figure_path}{Path to which figure is saved}

\item{plot_name}{ID of analysis for plot naming}

\item{color_scale}{Color scale to be applied}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Plot comparing V,J usage across repertoires
}
\examples{
vj_across_reps(list_repertoires,rep_category,chain="H",freq_threshold=0.01,unique_rows=TRUE,size_x_axis=5,size_y_axis=5,figure_path,plot_name="",color_scale)

}
