% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart_isotype_usage.R
\name{pie_chart_isotype_usage}
\alias{pie_chart_isotype_usage}
\title{Sideways barplot}
\usage{
pie_chart_isotype_usage(
  list_repertoires,
  chain = "H",
  order_ids,
  sample_name,
  text_size = 5,
  width_height = c(x = 15, y = 15),
  figure_path = "Ehling_figs",
  color_scale_category = c("#66c2a5", "#fc8d62", "#8da0cb", "#e78ac3", "#a6d854",
    "#ffd92f")
)
}
\arguments{
\item{list_repertoires}{Patient Metadata.}

\item{chain}{Heavy or Light chain}

\item{order_ids}{order in which pie charts should be shown}

\item{sample_name}{ID for plot name}

\item{text_size}{Size of text of pie labels}

\item{width_height}{width-height of full plot.}

\item{color_scale_category}{color scale.}

\item{plot_directory}{Path to which figures should be saved.}
}
\value{
TRUE (plots saved as pdfs into path_figures)
}
\description{
Sideways barplot
}
\examples{
pie_chart_isotype_usage(list_repertoires,chain="H",order_ids,sample_name)
}
